+++ cot/src/error.rs
@@ -1,459 +1,6 @@
 pub(crate) mod backtrace;
+pub(crate) mod error_impl;
 pub mod handler;
 pub mod method_not_allowed;
 pub mod not_found;
 pub mod uncaught_panic;
-
-use std::error::Error as StdError;
-use std::fmt::Display;
-
-use derive_more::with_trait::Debug;
-use thiserror::Error;
-
-use crate::StatusCode;
-// Need to rename Backtrace to CotBacktrace, because otherwise it triggers special behavior
-// in the thiserror library
-use crate::error::backtrace::{__cot_create_backtrace, Backtrace as CotBacktrace};
-
-/// An error that can occur while using Cot.
-pub struct Error {
-    inner: Box<ErrorImpl>,
-}
-
-impl Error {
-    #[must_use]
-    pub(crate) fn from_kind(inner: ErrorKind) -> Self {
-        let inner = ErrorImpl {
-            kind: inner,
-            backtrace: __cot_create_backtrace(),
-        };
-        Self {
-            inner: Box::new(inner),
-        }
-    }
-
-    /// Create a new error with a custom error message or error type.
-    ///
-    /// The error will be associated with a 500 Internal Server Error
-    /// status code, which is the default for unexpected errors.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use cot::Error;
-    ///
-    /// let error = Error::internal("An error occurred");
-    /// let error = Error::internal(std::io::Error::new(
-    ///     std::io::ErrorKind::Other,
-    ///     "An error occurred",
-    /// ));
-    /// ```
-    #[must_use]
-    pub fn internal<E>(error: E) -> Self
-    where
-        E: Into<Box<dyn StdError + Send + Sync + 'static>>,
-    {
-        Self::with_status(error, StatusCode::INTERNAL_SERVER_ERROR)
-    }
-
-    /// Create a new error with a custom error message or error type and a
-    /// specific HTTP status code.
-    ///
-    /// This method allows you to create an error with a custom status code,
-    /// which will be returned in the HTTP response. This is useful when you
-    /// want to return specific HTTP status codes like 400 Bad Request, 403
-    /// Forbidden, etc.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use cot::{Error, StatusCode};
-    ///
-    /// // Create a 400 Bad Request error
-    /// let error = Error::with_status("Invalid input", StatusCode::BAD_REQUEST);
-    ///
-    /// // Create a 403 Forbidden error
-    /// let error = Error::with_status("Access denied", StatusCode::FORBIDDEN);
-    /// ```
-    #[must_use]
-    pub fn with_status<E>(error: E, status_code: StatusCode) -> Self
-    where
-        E: Into<Box<dyn StdError + Send + Sync + 'static>>,
-    {
-        Self::from_kind(ErrorKind::Custom {
-            inner: error.into(),
-            status_code,
-        })
-    }
-
-    /// Create a new admin panel error with a custom error message or error
-    /// type.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use cot::Error;
-    ///
-    /// let error = Error::admin("An error occurred");
-    /// let error = Error::admin(std::io::Error::new(
-    ///     std::io::ErrorKind::Other,
-    ///     "An error occurred",
-    /// ));
-    /// ```
-    pub fn admin<E>(error: E) -> Self
-    where
-        E: Into<Box<dyn StdError + Send + Sync + 'static>>,
-    {
-        Self::from_kind(ErrorKind::AdminError(error.into()))
-    }
-
-    /// Create a new "404 Not Found" error without a message.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use cot::Error;
-    ///
-    /// let error = Error::not_found();
-    /// ```
-    #[must_use]
-    #[deprecated(
-        note = "Use `cot::Error::from(cot::error::not_found::NotFound::new())` instead",
-        since = "0.4.0"
-    )]
-    pub fn not_found() -> Self {
-        Self::from(not_found::NotFound::new())
-    }
-
-    /// Create a new "404 Not Found" error with a message.
-    ///
-    /// Note that the message is only displayed when Cot's debug mode is
-    /// enabled. It will not be exposed to the user in production.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use cot::Error;
-    ///
-    /// let id = 123;
-    /// let error = Error::not_found_message(format!("User with id={id} not found"));
-    /// ```
-    #[must_use]
-    #[deprecated(
-        note = "Use `cot::Error::from(cot::error::not_found::NotFound::with_message())` instead",
-        since = "0.4.0"
-    )]
-    pub fn not_found_message(message: String) -> Self {
-        Self::from(not_found::NotFound::with_message(message))
-    }
-
-    /// Returns the HTTP status code associated with this error.
-    ///
-    /// This method returns the appropriate HTTP status code that should be
-    /// sent in the response when this error occurs.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use cot::{Error, StatusCode};
-    ///
-    /// let error = Error::internal("Something went wrong");
-    /// assert_eq!(error.status_code(), StatusCode::INTERNAL_SERVER_ERROR);
-    ///
-    /// let error = Error::with_status("Bad request", StatusCode::BAD_REQUEST);
-    /// assert_eq!(error.status_code(), StatusCode::BAD_REQUEST);
-    /// ```
-    #[must_use]
-    pub fn status_code(&self) -> StatusCode {
-        self.inner.kind.status_code()
-    }
-
-    #[must_use]
-    pub(crate) fn kind(&self) -> &ErrorKind {
-        &self.inner.kind
-    }
-
-    #[must_use]
-    pub(crate) fn backtrace(&self) -> &CotBacktrace {
-        &self.inner.backtrace
-    }
-
-    /// If the error is a custom error, returns a reference to the inner
-    /// `cot::Error`, if any (recursively). If the error is not a custom
-    /// error, returns a reference to itself.
-    #[must_use]
-    pub fn inner(&self) -> Option<&Self> {
-        match &self.kind() {
-            ErrorKind::Custom { .. } | ErrorKind::MiddlewareWrapped { .. } => {
-                let mut error = self as &(dyn StdError + 'static);
-                while let Some(inner) = error.source() {
-                    if let Some(error) = inner.downcast_ref::<Self>() {
-                        return Some(error);
-                    }
-                    error = inner;
-                }
-                None
-            }
-            _ => Some(self),
-        }
-    }
-}
-
-impl Debug for Error {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        Debug::fmt(&self.inner, f)
-    }
-}
-
-impl Display for Error {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        Display::fmt(self.kind(), f)
-    }
-}
-
-impl StdError for Error {
-    fn source(&self) -> Option<&(dyn StdError + 'static)> {
-        self.kind().source()
-    }
-}
-
-impl From<ErrorKind> for Error {
-    fn from(value: ErrorKind) -> Self {
-        Self::from_kind(value)
-    }
-}
-
-#[derive(Debug)]
-struct ErrorImpl {
-    pub(crate) kind: ErrorKind,
-    #[debug(skip)]
-    backtrace: CotBacktrace,
-}
-
-macro_rules! impl_error_from_repr {
-    ($ty:ty) => {
-        impl From<$ty> for Error {
-            fn from(value: $ty) -> Self {
-                Error::from(ErrorKind::from(value))
-            }
-        }
-    };
-}
-
-impl From<Error> for askama::Error {
-    fn from(value: Error) -> Self {
-        askama::Error::Custom(Box::new(value))
-    }
-}
-
-impl_error_from_repr!(toml::de::Error);
-impl_error_from_repr!(askama::Error);
-impl_error_from_repr!(not_found::NotFound);
-impl_error_from_repr!(uncaught_panic::UncaughtPanic);
-impl_error_from_repr!(method_not_allowed::MethodNotAllowed);
-impl_error_from_repr!(crate::router::path::ReverseError);
-#[cfg(feature = "db")]
-impl_error_from_repr!(crate::db::DatabaseError);
-impl_error_from_repr!(tower_sessions::session::Error);
-impl_error_from_repr!(crate::form::FormError);
-impl_error_from_repr!(crate::form::FormFieldValueError);
-impl_error_from_repr!(crate::auth::AuthError);
-impl_error_from_repr!(crate::request::PathParamsDeserializerError);
-impl_error_from_repr!(crate::request::extractors::StaticFilesGetError);
-
-#[derive(Debug, Error)]
-#[non_exhaustive]
-pub(crate) enum ErrorKind {
-    /// A custom user error occurred.
-    #[error("{inner}")]
-    Custom {
-        #[source]
-        inner: Box<dyn StdError + Send + Sync>,
-        status_code: StatusCode,
-    },
-    /// An error occurred while trying to load the config.
-    #[error("Could not read the config file at `{config}` or `config/{config}.toml`")]
-    LoadConfig {
-        config: String,
-        source: std::io::Error,
-    },
-    /// An error occurred while trying to parse the config.
-    #[error("Could not parse the config: {source}")]
-    ParseConfig {
-        #[from]
-        source: toml::de::Error,
-    },
-    /// An error occurred while trying to start the server.
-    #[error("Could not start server: {source}")]
-    StartServer { source: std::io::Error },
-    /// An error occurred while trying to collect static files into a directory.
-    #[error("Could not collect static files: {source}")]
-    CollectStatic { source: std::io::Error },
-    /// An error occurred while trying to read the request body.
-    #[error("Could not retrieve request body: {source}")]
-    ReadRequestBody {
-        #[source]
-        source: Box<dyn StdError + Send + Sync>,
-    },
-    /// The request body had an invalid `Content-Type` header.
-    #[error("Invalid content type; expected `{expected}`, found `{actual}`")]
-    InvalidContentType {
-        expected: &'static str,
-        actual: String,
-    },
-    /// The request does not contain a form.
-    #[error(
-        "Request does not contain a form (expected `application/x-www-form-urlencoded` or \
-        `multipart/form-data` content type, or a GET or HEAD request)"
-    )]
-    ExpectedForm,
-    /// Could not find a route for the request.
-    #[error("{0}")]
-    NotFound(#[from] not_found::NotFound),
-    #[error("An unexpected error occurred")]
-    UncaughtPanic(#[from] uncaught_panic::UncaughtPanic),
-    #[error("{0}")]
-    MethodNotAllowed(#[from] method_not_allowed::MethodNotAllowed),
-    /// Could not create a response object.
-    #[error("Could not create a response object: {0}")]
-    ResponseBuilder(#[from] http::Error),
-    /// `reverse` was called on a route that does not exist.
-    #[error("Failed to reverse route `{view_name}` due to view not existing")]
-    NoViewToReverse {
-        app_name: Option<String>,
-        view_name: String,
-    },
-    /// An error occurred while trying to reverse a route (e.g. due to missing
-    /// parameters).
-    #[error("Failed to reverse route: {0}")]
-    ReverseRoute(#[from] crate::router::path::ReverseError),
-    /// An error occurred while trying to render a template.
-    #[error("Failed to render template: {0}")]
-    TemplateRender(#[from] askama::Error),
-    /// An error occurred while communicating with the database.
-    #[error("Database error: {0}")]
-    #[cfg(feature = "db")]
-    Database(#[from] crate::db::DatabaseError),
-    /// An error occurred while accessing the session object.
-    #[error("Error while accessing the session object")]
-    SessionAccess(#[from] tower_sessions::session::Error),
-    /// An error occurred while parsing a form.
-    #[error("Failed to process a form: {0}")]
-    Form(#[from] crate::form::FormError),
-    /// An error occurred while trying to retrieve the value of a form field.
-    #[error("Failed to retrieve the value of a form field: {0}")]
-    FormFieldValueError(#[from] crate::form::FormFieldValueError),
-    /// An error occurred while trying to authenticate a user.
-    #[error("Failed to authenticate user: {0}")]
-    Authentication(#[from] crate::auth::AuthError),
-    /// An error occurred while trying to serialize or deserialize JSON.
-    #[error("JSON error: {0}")]
-    #[cfg(feature = "json")]
-    Json(serde_path_to_error::Error<serde_json::Error>),
-    /// An error occurred inside a middleware-wrapped view.
-    #[error(transparent)]
-    MiddlewareWrapped {
-        source: Box<dyn StdError + Send + Sync>,
-    },
-    /// An error occurred while trying to parse path parameters.
-    #[error("Could not parse path parameters: {0}")]
-    PathParametersParse(#[from] crate::request::PathParamsDeserializerError),
-    /// An error occurred while trying to parse query parameters.
-    #[error("Could not parse query parameters: {0}")]
-    QueryParametersParse(serde_path_to_error::Error<serde::de::value::Error>),
-    /// An error occurred in an [`AdminModel`](crate::admin::AdminModel).
-    #[error("Admin error: {0}")]
-    AdminError(#[source] Box<dyn StdError + Send + Sync>),
-    /// An error occurred while getting a URL for a static files.
-    #[error("Could not get URL for a static file: {0}")]
-    StaticFilesGetError(#[from] crate::request::extractors::StaticFilesGetError),
-}
-
-impl ErrorKind {
-    fn status_code(&self) -> StatusCode {
-        // TODO check if works with middleware
-        match self {
-            // 500 Internal Server Error
-            ErrorKind::AdminError(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::Authentication(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::CollectStatic { .. } => StatusCode::INTERNAL_SERVER_ERROR,
-            #[cfg(feature = "db")]
-            ErrorKind::Database(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::Form(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::FormFieldValueError(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            #[cfg(feature = "json")]
-            ErrorKind::Json(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::LoadConfig { .. } => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::MiddlewareWrapped { .. } => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::NoViewToReverse { .. } => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::ParseConfig { .. } => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::ResponseBuilder(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::ReverseRoute(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::SessionAccess(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::StartServer { .. } => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::StaticFilesGetError(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::TemplateRender(_) => StatusCode::INTERNAL_SERVER_ERROR,
-            ErrorKind::UncaughtPanic { .. } => StatusCode::INTERNAL_SERVER_ERROR,
-            // 400 Bad Request
-            ErrorKind::ExpectedForm => StatusCode::BAD_REQUEST,
-            ErrorKind::InvalidContentType { .. } => StatusCode::BAD_REQUEST,
-            ErrorKind::PathParametersParse(_) => StatusCode::BAD_REQUEST,
-            ErrorKind::QueryParametersParse(_) => StatusCode::BAD_REQUEST,
-            ErrorKind::ReadRequestBody { .. } => StatusCode::BAD_REQUEST,
-            // Other
-            ErrorKind::MethodNotAllowed(_) => StatusCode::METHOD_NOT_ALLOWED,
-            ErrorKind::NotFound { .. } => StatusCode::NOT_FOUND,
-            // Custom
-            ErrorKind::Custom { status_code, .. } => *status_code,
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::io;
-
-    use super::*;
-
-    #[test]
-    fn test_error_new() {
-        let inner = ErrorKind::StartServer {
-            source: io::Error::other("server error"),
-        };
-
-        let error = Error::from_kind(inner);
-
-        assert!(StdError::source(&error).is_some());
-    }
-
-    #[test]
-    fn test_error_display() {
-        let inner = ErrorKind::InvalidContentType {
-            expected: "application/json",
-            actual: "text/html".to_string(),
-        };
-        let error = Error::from_kind(inner);
-
-        let display = format!("{error}");
-
-        assert_eq!(
-            display,
-            "Invalid content type; expected `application/json`, found `text/html`"
-        );
-    }
-
-    #[test]
-    fn test_error_from_repr() {
-        let inner = ErrorKind::NoViewToReverse {
-            app_name: None,
-            view_name: "home".to_string(),
-        };
-
-        let error: Error = inner.into();
-
-        assert_eq!(
-            format!("{error}"),
-            "Failed to reverse route `home` due to view not existing"
-        );
-    }
-}
